Updated documentation and refactored structure
