+++ src/main/java/md2k/mCerebrum/CSVDataPoint.java
@@ -1,18 +1,20 @@
-package md2k.mCerebrum.cStress.library;/*
- * Copyright (c) 2015, The University of Memphis, MD2K Center 
+package md2k.mCerebrum;
+
+/*
+ * Copyright (c) 2015, The University of Memphis, MD2K Center
  * - Timothy Hnat <twhnat@memphis.edu>
  * All rights reserved.
- * 
+ *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * * Redistributions of source code must retain the above copyright notice, this
  *   list of conditions and the following disclaimer.
- * 
+ *
  * * Redistributions in binary form must reproduce the above copyright notice,
  *   this list of conditions and the following disclaimer in the documentation
  *   and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -25,10 +27,36 @@ package md2k.mCerebrum.cStress.library;/*
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-public interface DataStreamInterface {
+/**
+ * DataPoint object to handle CSV AutoSense data files
+ */
+public class CSVDataPoint implements Comparable {
+    public double value;
+    public long timestamp;
+    public int channel;
 
-    void persist(String filename);
 
-    void reset();
+    /**
+     * @param channel   AutoSense channel which determine a specific sensor
+     * @param timestamp Timestamp in milliseconds since Jan 1st, 1970
+     * @param value     Sensor value
+     */
+    public CSVDataPoint(int channel, long timestamp, double value) {
+        this.channel = channel;
+        this.value = value;
+        this.timestamp = timestamp;
+    }
+
+    @Override
+    public int compareTo(Object obj) {
+        CSVDataPoint o = (CSVDataPoint) obj;
+        if (this.value == o.value && this.timestamp == o.timestamp && this.channel == o.channel) {
+            return 0; //They are the same
+        } else if (this.timestamp < o.timestamp) {
+            return -1; //"this" is before "o"
+        } else {
+            return 1; //"this" is after "o"
+        }
+    }
 
 }
