+++ src/main/java/md2k/mCerebrum/cStress/Autosense/AUTOSENSE_PACKET.java
@@ -34,7 +34,7 @@ public class AUTOSENSE_PACKET {
 
     /**
      * Timestamp in milliseconds.
-     * <p/>
+     *
      * Packet timestamp as recorded by the Android phone
      */
     public long timestamp;
