+++ src/main/java/md2k/mCerebrum/cStress/Library/DataQuality/AutoSense/ECGQualityCalculation.java
@@ -162,6 +162,12 @@ public class ECGQualityCalculation {
         return AUTOSENSE.QUALITY_GOOD;
     }
 
+    /**
+     * Interface routine between the old Java AutoSense quality calculation and stream-processor
+     * @param ecg Input ECG DataPoint list
+     * @param windowSize The size of a window in milliseconds
+     * @return DataPoint list of qualities
+     */
     public List<DataPoint> computeQuality(List<DataPoint> ecg, long windowSize) {
 
         List<DataPoint[]> windowedECG = Time.window(ecg, windowSize);
